Move some functions to powerline.lib.dict
Moved functions: `merge*` from `powerline.lib` and `mergeargs` from
`powerline.commands.main`.
